test(flip-api): dedupe log-policy helper + pin static detail strings
Address review feedback on PR #426 from @garciadias.

* **Extract shared helper.** The ~25-line
  `_assert_logs_have_no_presigned_url` (plus `_FAKE_SIGNED_URL`,
  `_S3_URL_WITH_QUERY`, `_FORBIDDEN_TOKENS`) was copy-pasted across
  `test_s3_client.py`, `test_presigned_url_for_upload.py`, and
  `test_retrieve_federated_results.py`. Moved to
  `flip-api/tests/unit/_log_policy.py` so a new forbidden token or
  pattern lands in one place and all three consumers stay in sync.
  Used a regular module rather than `conftest.py` — the symbols are
  imported by name (not pytest fixtures), and pytest's docs steer
  `conftest.py` toward fixture/hook definitions.

* **Converge the two narrow assertion loops.** Two tests in
  `test_presigned_url_for_upload.py` and `test_retrieve_federated_results.py`
  had stale narrow loops checking only `record.getMessage()` against
  `_FORBIDDEN_TOKENS`, with comments that no longer reflected reality
  (the routes use `logger.error`, not `logger.exception`). Switched
  them to the shared helper — which also inspects `record.exc_info`
  and `record.exc_text` — so the policy is enforced uniformly.

* **Pin exact static detail strings.** Tightened
  `assert "Error deleting" in detail` → `assert detail == "Error deleting file"`
  in `test_delete_file.py:103`, same treatment in `test_download_file.py:166`
  (`"Error downloading file"`). The substring match would pass even
  if a future regression re-introduced `f"Error deleting: {exc}"`.

Retest: 921 unit tests pass; ruff + mypy clean.
